api/views_updated.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework import status
import requests
from django.conf import settings
from datetime import datetime


@api_view(['GET'])
def commodities_view(request):
    """
    Fetch real-time commodity data from AlphaVantage API
    Returns processed data for chart visualization
    """
    try:
        api_key = settings.ALPHAVANTAGE_API_KEY
        url = f"https://www.alphavantage.co/query?function=ALL_COMMODITIES&interval=monthly&apikey={api_key}"
        
        response = requests.get(url, timeout=10)
        response.raise_for_status()
        
        api_data = response.json()
        
        # Check if we got valid data
        if 'data' not in api_data:
            raise ValueError("Invalid API response")
        
        # Process the commodity data for chart
        chart_data = []
        commodities_list = api_data['data']
        
        # Group data by date and calculate average or use specific commodity
        # For this demo, we'll use the first commodity's data points
        date_values = {}
        
        for item in commodities_list:
            try:
                date_str = item.get('date', '')
                value_str = item.get('value', '0')
                
                # Parse the value (remove any non-numeric characters except decimal point and minus)
                value = float(value_str.replace(',', ''))
                
                if date_str not in date_values:
                    date_values[date_str] = []
                date_values[date_str].append(value)
            except (ValueError, KeyError, AttributeError):
                continue
        
        # Convert to chart format (take last 21 data points)
        sorted_dates = sorted(date_values.keys(), reverse=True)[:21]
        sorted_dates.reverse()  # Show oldest to newest
        
        for date_str in sorted_dates:
            try:
                # Calculate average value for the date
                avg_value = sum(date_values[date_str]) / len(date_values[date_str])
                
                # Format date for display
                dt = datetime.strptime(date_str, '%Y-%m-%d')
                formatted_date = dt.strftime('%b %d')
                
                chart_data.append({
                    'name': formatted_date,
                    'value': round(avg_value, 2)
                })
            except (ValueError, KeyError):
                continue
        
        # If we don't have enough data points, use sample data
        if len(chart_data) < 10:
            chart_data = get_sample_data()
        
        return Response({
            'success': True,
            'data': chart_data,
            'source': 'alphavantage_api',
            'total_records': len(commodities_list)
        })
        
    except Exception as e:
        # Return sample data if API fails
        return Response({
            'success': True,
            'data': get_sample_data(),
            'source': 'sample_data',
            'message': f'Using sample data: {str(e)}'
        })


@api_view(['GET'])
def users_view(request):
    """
    Return real-time user/broker data matching the dashboard format
    In a real application, this would fetch from a database
    """
    users_data = [
        {
            "id": 1,
            "broker": "Zerodha (DU000004)",
            "no_of_active_positions": 1,
            "available_capital": "₹ 1.54 Cr",
            "total_deployed_strategies": 3,
            "active_strategies": 1,
            "status": "Active",
            "current_pnl": "₹ 50.02 K",
            "required_capital": "₹ 50.02 K"
        },
        {
            "id": 2,
            "broker": "Angel One (MNBN1026)",
            "no_of_active_positions": 2,
            "available_capital": "₹ 2.50 K",
            "total_deployed_strategies": 2,
            "active_strategies": 2,
            "status": "Active",
            "current_pnl": "₹ 60.02 K",
            "required_capital": "₹ 60.02 K"
        },
        {
            "id": 3,
            "broker": "Finvasia (FA189009)",
            "no_of_active_positions": 0,
            "available_capital": "₹ 50.02 K",
            "total_deployed_strategies": 0,
            "active_strategies": 0,
            "status": "Pending",
            "current_pnl": "₹ 0.00",
            "required_capital": "₹ 0.00"
        }
    ]
    
    return Response({
        'success': True,
        'data': users_data,
        'timestamp': datetime.now().isoformat()
    })


def get_sample_data():
    """
    Returns sample data for chart when API is unavailable
    """
    return [
        {"name": "Dec 1", "value": 10},
        {"name": "Dec 2", "value": -35},
        {"name": "Dec 3", "value": 20},
        {"name": "Dec 4", "value": 15},
        {"name": "Dec 5", "value": 25},
        {"name": "Dec 6", "value": 35},
        {"name": "Dec 7", "value": 30},
        {"name": "Dec 8", "value": 5},
        {"name": "Dec 9", "value": -40},
        {"name": "Dec 10", "value": 15},
        {"name": "Dec 11", "value": 10},
        {"name": "Dec 12", "value": 25},
        {"name": "Dec 13", "value": 5},
        {"name": "Dec 14", "value": -20},
        {"name": "Dec 15", "value": -25},
        {"name": "Dec 16", "value": -10},
        {"name": "Dec 17", "value": -30},
        {"name": "Dec 18", "value": 5},
        {"name": "Dec 19", "value": 35},
        {"name": "Dec 20", "value": 20},
        {"name": "Dec 21", "value": 20}
    ]
